move dialog back, change exposed names ;(
stage 1
